requirements.txt
imageio==2.9.0
ipdb==0.12.3
joblib==1.0.1
lz4==3.1.3
matplotlib==3.1.3
numpy==1.20.3
open3d==0.10.0.0
opencv_contrib_python==3.4.2.17
path.py==12.5.0
Pillow==8.4.0
pyhocon==0.3.58
pykdtree==1.3.4
requests==2.26.0
scipy==1.7.1
tensorboardX==2.4.1
torch==1.5.0
torchvision==0.6.0
